Update position handling for u32

+++ crates/server/src/util/mod.rs
@@ -7,13 +7,13 @@ pub(crate) mod node {
         /// Creates an lsp position from the starting position of a tree-sitter node.
         pub(crate) fn start(node: &tree_sitter::Node) -> tower_lsp::lsp_types::Position {
             let tree_sitter::Point { row, column } = node.start_position();
-            tower_lsp::lsp_types::Position::new(row as u64, column as u64)
+            tower_lsp::lsp_types::Position::new(row as u32, column as u32)
         }
 
         /// Creates an lsp position from the ending position of a tree-sitter node.
         pub(crate) fn end(node: &tree_sitter::Node) -> tower_lsp::lsp_types::Position {
             let tree_sitter::Point { row, column } = node.end_position();
-            tower_lsp::lsp_types::Position::new(row as u64, column as u64)
+            tower_lsp::lsp_types::Position::new(row as u32, column as u32)
         }
     }
